HUGE MAJOR PROGRESS
NOT ONLY WORKING BUTTON BUT NOW FLASK COMMUNICATES BETWEEN TWO. Need to initalize the python script when created but if server is running it will communicate and print to log apple share price thru the twelve data api. so excited and proud typing this to myself but idgas cause im so fuckiing excited. BIG BIG BIG BIG